requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.19.0
matplotlib>=3.3.0
pandas>=1.1.0
Pillow>=8.0.0
tqdm>=4.48.0
opencv-python>=4.4.0
tensorboard>=2.4.0
albumentations>=0.5.2